+++ xayn-ai/src/ranker/context.rs
@@ -1,4 +1,4 @@
-use crate::{data::document_data::CoiComponent, ranker::util::Id};
+use crate::{data::document_data::CoiComponent, ranker::utils::Id};
 
 /// The context used to calculate a document's score.
 /// https://xainag.atlassian.net/wiki/spaces/M2D/pages/770670607/Production+AI+Workflow#3.2-Context-calculations.
@@ -37,7 +37,7 @@ impl Context {
 mod tests {
     use crate::{
         data::document_data::CoiComponent,
-        ranker::{context::Context, util::Id},
+        ranker::{context::Context, utils::Id},
         CoiId,
     };
     use test_utils::assert_approx_eq;
